refactor: changing structure
