+++ tasker_service/app/core/__init__.py
@@ -1,0 +1,6 @@
+import asyncio
+from typing import Dict
+
+tasks_db: Dict[str, dict] = {}
+
+task_queue: asyncio.Queue = asyncio.Queue()
